requirements.txt
mysql-connector-python==8.0.27
pandas==1.3.5
streamlit==1.3.1
st-material-table==0.0.4
plotly==5.5.0
xlsxwriter==3.0.2